Moved boundary_mesh_constructor to plottting
